% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_gps.R
\name{cor2cov}
\alias{cor2cov}
\title{cor2cov function}
\usage{
cor2cov(ld.cor.mat, v.vec)
}
\arguments{
\item{ld.cor.mat}{a LD correlation matrix}

\item{v.vec}{a vector of v statistic}
}
\value{
the approximated X'X from summary statistics and LD.
}
\description{
cor2cov function
}
