% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_lasso_sum_ess_gps}
\alias{fast_lasso_sum_ess_gps}
\title{core function for fast coordinate descent of GPS loss function}
\usage{
fast_lasso_sum_ess_gps(
  uVec,
  xTx,
  n,
  group,
  lambda,
  alpha,
  eta,
  bccVec,
  initVec,
  maxIter = 100L
)
}
\arguments{
\item{uVec}{u statistics (x'y)}

\item{xTx}{x_transpose_x matrix calculated from reference panel}

\item{n}{sample size;}

\item{group}{1-based group indicator that indicates which alpha and lambda to use}

\item{lambda}{l2 penalty parameter options, match unique group id}

\item{alpha}{l1 penalty parameter options, match unique group id}

\item{eta}{l2 penalty parameter for deviation from prior (correlated) effect size estimates}

\item{bccVec}{weight of PRS for a correlated trait to serve as a prior}

\item{initVec}{initial value of beta (joint effect)}

\item{maxIter}{maximum number of iterations}
}
\value{
a list where \code{beta} is the fitted regression coefficient vector, and \code{iteration} is the actual iteration.
}
\description{
core function for fast coordinate descent of GPS loss function
}
