% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{gps}
\alias{gps}
\title{gps: GPS is a penalized regression based method that can integrate large scale case-control study
and electronic health record (EHR) based biobanks to construct polygenic risk score models for disease
progression risk from preclinical stage with improved accuracy.}
\description{
_PACKAGE
}
