% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.sum.ess.R
\name{gps_fit_global}
\alias{gps_fit_global}
\title{fit global}
\usage{
gps_fit_global(
  beta_est,
  se_est,
  ld_cor_list,
  sample_size,
  anno_group,
  weight_prior,
  lambda,
  pi,
  phi,
  eta,
  init_vec = NULL,
  max_iter,
  snp_id
)
}
\arguments{
\item{b.vec}{genetic effect estimates}
}
\value{
a vector
}
\description{
fit global
}
