% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.sum.ess.R
\name{gps_fit_single}
\alias{gps_fit_single}
\title{fit a locus}
\usage{
gps_fit_single(
  beta_est,
  se_est,
  ld_cor_mt,
  sample_size,
  anno_group,
  weight_prior,
  lambda,
  pi,
  phi,
  eta,
  init_vec = NULL,
  max_iter = 1000,
  snp_id = NULL
)
}
\arguments{
\item{b.vec}{genetic effect estimates}
}
\value{
a vector
}
\description{
fit a locus
}
