% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.sum.ess.R
\name{gps_grid}
\alias{gps_grid}
\title{fit grid}
\usage{
gps_grid(
  min_ratio = 0.01,
  nlambda = 30,
  l1_ratio = c(0, 0.25, 0.5, 0.75, 1),
  eta_vec,
  beta_est,
  se_est,
  ld_cor_list,
  sample_size,
  weight_prior,
  init_vec = NULL,
  max_iter = 1000
)
}
\arguments{
\item{b.vec}{genetic effect estimates}
}
\value{
a vector
}
\description{
fit grid
}
