% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_lasso_sum_ess_gps}
\alias{fast_lasso_sum_ess_gps}
\title{Rewrite lasso.sum.ess}
\usage{
fast_lasso_sum_ess_gps(
  uVec,
  xTx,
  n,
  group,
  lambda,
  alpha,
  eta,
  bccVec,
  initVec,
  maxIter = 100L
)
}
\arguments{
\item{n}{sample size;}

\item{group}{1-based group indicator}

\item{lambda}{l1 penalty parameter}

\item{alpha}{l2 penalty parameter}

\item{initVec}{initial value of beta}

\item{maxIter}{maximum number of iterations}

\item{bVec}{genetic effect estimates}

\item{sVec}{standard error of genetic effect estimates;}

\item{r2Mat}{residual errors;}
}
\value{
a list where `beta` is the fitted regression coefficient vector, and `iteration` is the actual iteration.
}
\description{
Rewrite lasso.sum.ess
}
