% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_gps.R
\name{gps_grid}
\alias{gps_grid}
\title{fit grid}
\usage{
gps_grid(
  min_ratio = 0.01,
  nlambda = 30,
  l1_ratio = c(0, 0.25, 0.5, 0.75, 1),
  eta_vec,
  beta_est,
  se_est,
  ld_cor_list,
  sample_size,
  weight_prior,
  init_vec = NULL,
  max_iter = 1000
)
}
\arguments{
\item{min_ratio}{the ratio between the minimum lambda (hyperparameter for shrinkage) and the maximum lambda}

\item{nlambda}{number of lambda values that are evenly spaced beween min of lambda and max of lambda}

\item{l1_ratio}{the contribution of l1 penalty for shrinkage.}

\item{eta_vec}{a vector of eta values to be used. eta controls the contribution of the prior PRS model weights.}

\item{beta_est}{a vector of marginal effect size estimates (assuming standardized genotype and phenotype)}

\item{se_est}{a vector of standard error estimates of marginal effect size}

\item{ld_cor_list}{a list of LD correlation matrix, each element is a LD correlation matrix of a LD block}

\item{sample_size}{a scalar of sample size}

\item{weight_prior}{a vector of the PRS weights of a correlated traits (e.g. PRS of the case-control phenotype).
The order of the elements should be the same as beta_est and se_est. If not available, please fill with NAs.}

\item{init_vec}{initial values used for coordinate descent algorithm.}

\item{max_iter}{num of maximum iterations used in coordinate descent}
}
\value{
List of two elements: 1. a grid of joint effect size estimates. Each column corresponds to a specific combination of hyperparameters
2. a dataframe containing corresponding combinations of hyperparameters
}
\description{
fit grid
}
